connected events to the proper route, rename protected client to public (#53)
