+++ src/components/event-listing/EventListing.tsx
@@ -44,8 +44,7 @@ const Info = styled.div`
 const EventListing: React.FC<EventProps> = ({
   thumbnail,
   title,
-  date,
-  description,
+  details,
 }) => {
   const defaultImg =
     'https://today.tamu.edu/wp-content/uploads/2019/03/GettyImages-184621282.jpg';
@@ -56,11 +55,11 @@ const EventListing: React.FC<EventProps> = ({
         <Thumbnail src={thumbnail || defaultImg} />
         <Info>
           <Title level={3}>{title}</Title>
-          <Text strong>{dateFormat(date, 'longDate')}</Text>
+          <Text strong>{dateFormat(details.start, 'longDate', true)}</Text>
           <br />
-          <Text strong>{dateFormat(date, 'shortTime')}</Text>
+          <Text strong>{dateFormat(details.start, 'shortTime', true)}</Text>
           <ThinDivider />
-          <Paragraph ellipsis={{ rows: 5 }}>{description}</Paragraph>
+          <Paragraph ellipsis={{ rows: 5 }}>{details.description}</Paragraph>
           <GreenButton>Register</GreenButton>
         </Info>
       </CardContent>
