+++ src/containers/upcoming-events/ducks/types.ts
@@ -2,7 +2,7 @@ import { AsyncRequest } from '../../../utils/asyncRequest';
 import { C4CState } from '../../../store';
 import { ThunkAction } from 'redux-thunk';
 import { EventsActions } from './actions';
-import { ApiExtraArgs } from '../../../api/protectedApiClient';
+import { ApiExtraArgs } from '../../../api/publicApiClient';
 
 export interface EventsReducerState {
   readonly upcomingEvents: AsyncRequest<EventProps[], any>;
@@ -16,9 +16,16 @@ export type EventsThunkAction<R> = ThunkAction<
 >;
 
 export interface EventProps {
+  id: number;
   title: string;
-  date: Date;
-  description: string;
-  otherNotes?: string;
+  spotsAvailable: number;
+  capacity: number;
   thumbnail?: string;
+  details: {
+    description: string;
+    location: string;
+    start: Date;
+    end: Date;
+  }
+  price: number;
 }
